+++ unit-tests/3d/test_mpi_parcel_init_3d.f90
@@ -79,7 +79,7 @@ program test_mpi_parcel_init_3d
         enddo
     enddo
 
-    call field_halo_swap(tbuoyg)
+    call field_halo_swap_scalar(tbuoyg)
 
     !---------------------------
     ! Generate parcel attribute: